#130 Split up integration tests and unit tests out of test_caiso.py
They now follow the directory conventions being established with #130.
